Move build variant card files
